copybot: Add filter CL option
The current behavior when subtrees are provided is to filter the CL
to the subtree.  In some cases, it's desirable to only take CL's
which affect the subtree, but still take the full content of the
repo when changes are made.  Add an option for resetting subtree
definitions after the list of CL's to downstream is found if the
config option is provided.

BUG=None
TEST=run copybot with no-filter-changes

Change-Id: I031e58e9974b96ed83a902a4d6d1f9ed77039562
Reviewed-on: https://chromium-review.googlesource.com/c/copybot/+/6324750
Tested-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jonathon Murphy <jpmurphy@google.com>