copybot: Add config file update
Infra team asked if we could be more on top of updating the origin
hashes to minimize the load on the bots. The goal of this change is to
have copybot update its own configuration after it has found the most
recent related change.

The caveat to this being that, we will be going back to an old behavior
that merging CL's out of order could cause missing changes. Since
copybot is doing the vast majority of the changes, and it is configured
to auto-submit CL's, this should be a reasonable trade-off.

BUG=None
TEST=./copybot.py --config
./config/coreboot-main-copybot-downstream.ini

Change-Id: Ic09a669f1953d34dd1d76c4502f38524388dcc8e
Reviewed-on: https://chromium-review.googlesource.com/c/copybot/+/6515237
Reviewed-by: Tim Bain <tbain@google.com>
Auto-Submit: Jon Murphy <jpmurphy@google.com>
Commit-Queue: Jon Murphy <jpmurphy@google.com>
Tested-by: Jon Murphy <jpmurphy@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>