copybot: handle matching commit on upstream and downstream
The logic for finding the last merged revision seems to somehow missing
the case where the two repositories have an actual matching commit, that
is same commit sha, not a cherry picked one with matching revid.

The logic was mostly implemented as this branch enters if

rev in upstream_hashes

and can calculate the index with

origin_revid or rev

but right now only triggers this condition based on the origin_revid.

Add the new condition to fix the issue.

BUG=none
TEST=./copybot.py --dry-run --config config/zephyr-cmsis_6-copybot-downstream.ini

Change-Id: Ib3c9e96d2d0c7dfe4a3b0eca0dbb42456482d57a
Reviewed-on: https://chromium-review.googlesource.com/c/copybot/+/6578572
Reviewed-by: Jon Murphy <jpmurphy@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Auto-Submit: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Jon Murphy <jpmurphy@google.com>