[kernel-cl-dispatcher] Search for pending changes in Android hosts
Copybot only looks for pending changes (in review) if the host is a
Git-on-Borg, which is checked by a regex match.

Without that, Copybot created a new CL on each run. This patch lists
Android hosts as eligible for the check, so that duplicate CLs are not
issued on each copybot run.

BUG=b:388434764
TEST=Ran command from the last copybot run and ensured it does not
create a new CL but rather acknowledges the pending one.

Change-Id: I14382c43ca0beb8297251b7cb91b60c3f12edc38
Reviewed-on: https://chromium-review.googlesource.com/c/copybot/+/7209371
Auto-Submit: Mateusz Urbańczyk <urbanczyk@google.com>
Commit-Queue: Jon Murphy <jpmurphy@google.com>
Reviewed-by: Jon Murphy <jpmurphy@google.com>
Reviewed-by: Jędrzej Ciupis <jciupis@google.com>
Tested-by: Mateusz Urbańczyk <urbanczyk@google.com>